+++ auth/services/services.go
@@ -61,7 +61,5 @@ type ContractNotary interface {
 
 	Configure() error
 
-	// Start any validator that needs to periodically update its database. Cancel the context to stop these processes.
-	Start(ctx context.Context)
 	core.Routable
 }