+++ test/node/server.go
@@ -24,6 +24,8 @@ import (
 	"net/http"
 	"os"
 	"path"
+	"path/filepath"
+	"runtime"
 	"sync"
 	"testing"
 	"time"
@@ -69,6 +71,17 @@ func StartServer(t *testing.T, configFunc ...func(httpServerURL string)) (string
 	t.Setenv("NUTS_NETWORK_GRPCADDR", grpcPort)
 	t.Setenv("NUTS_EVENTS_NATS_PORT", natsPort)
 	t.Setenv("NUTS_AUTH_PUBLICURL", httpServerURL)
+
+	// relative paths are defined from the location source 't', find absolute path of current function to fix this
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		t.Fatal("Unable to get the current filename")
+	}
+	dirname := filepath.Dir(filename)
+
+	t.Setenv("NUTS_TLS_CERTFILE", fmt.Sprintf("%s/../pki/certificate-and-key.pem", dirname))
+	t.Setenv("NUTS_TLS_CERTKEYFILE", fmt.Sprintf("%s/../pki/certificate-and-key.pem", dirname))
+	t.Setenv("NUTS_TLS_TRUSTSTOREFILE", fmt.Sprintf("%s/../pki/truststore.pem", dirname))
 
 	for _, fn := range configFunc {
 		fn(httpServerURL)